+++ src/main/java/com/kuna_backend/repositories/ShippingDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShippingDetailRepository extends JpaRepository <ShippingDetail, Integer> {
+public interface ShippingDetailRepository extends JpaRepository <ShippingDetail, Long> {
 }
