+++ src/main/java/com/kuna_backend/repositories/TokenRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TokenRepository extends JpaRepository <AuthenticationToken, Integer> {
+public interface TokenRepository extends JpaRepository <AuthenticationToken, Long> {
     AuthenticationToken findTokenByClient(Client client);
     AuthenticationToken findTokenByToken(String token);
 }
